+++ src/Component/index.tsx
@@ -3,5 +3,6 @@ import LoadingState from './LoadingState';
 import LoadingSpinner from './LoadingSpinner';
 import Toast from './Toast';
 import WeatherToast from './WeatherToast';
+import Tutorial from './Tutorial';
 
-export { Layout, LoadingState, LoadingSpinner, Toast, WeatherToast };
+export { Layout, LoadingState, LoadingSpinner, Toast, WeatherToast, Tutorial };
